requirements.txt
pandas
numpy
gradio
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma